Refactor gateway to parse gear from domin in app middleware
