+++ pkg/gateway/model/context.go
@@ -9,8 +9,8 @@ type contextKey string
 const contextKeyGatewayContext contextKey = "gateway-context"
 
 type Context struct {
-	App    App
-	Domain Domain
+	App  App
+	Gear Gear
 }
 
 func ContextWithGatewayContext(ctx context.Context, gatewayContext Context) context.Context {
